Finished `datalog_trans!()`-implementation
Note, however, that our idea with transitions wont really work lmao.

The problem is that the conditions aren't applied on the facts at creation time, but instead persistent at the inference time (the condition, that is). That makes it impossible to have lingering facts; not lingering rules.

Doesn't make a lot of sense. I know  the problem, though xD