Reworked transition postulations
Now, the persistance of postulations is over computed instances instead of rules. This is nice, and allows us to model the state system; however, currently, the conditionalism is not yet quantifying (or rather, the rule is, but us checking for produced values isn't). Will probably have to make a MUCH nicer interface to the existing quantification, then use that.